refactor(web): interview 삭제 로직 이동 및 성공시 콜백 함수 호출
